chore(ci): test Travis PR setup
This commit is just to trigger the "travis/pr" job and see what happens.
Let's see if this passes or not.

Apparently there's an issue with Chrome stable not working on Travis
with `dist: trusty` anymore, so will probably have to update `dist`.

Related: https://github.com/vaadin/vaadin-element-skeleton/pull/211