test: update visual test screenshots
Flex behaviour is now correct (also in old Firefox, Edge and IE11).
The updated Firefox screnshots now show an alignment issue, but that does
not seem to affect recent Firefox versions or even Edge/IE11 so it can
be ignored. (Would be good to be able to run these tests on recent FF.)