fix: too many value-change events
Small update to __formatDateTime() makes it so that __handleDateTimeChange() isn't executed too many times when value is being cleared. Previously when clearing the value it was being first set to "" then temprorarily set to string "T" and then reverted (as invalid) back to "". This triggered 3 value-change events when clearing the value.

Also removed a small block on unnecessary code.